# backend/tests/contract/openapi.json
{
 "components": {
  "schemas": {
...
   "Currency": {
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "code",
     "name"
    ],
    "title": "Currency",
    "type": "object"
   },
...
   "FIIDividendResponse": {
    "description": "One payment per share, as published.\n\n``event_type`` is what the fund called it — an ordinary distribution or an\namortization of capital. It is carried so that a reader meaning income can\ntell the two apart instead of adding them up.",
    "properties": {
     "event_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Event Type"
     },
     "ex_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ex Date"
     },
     "payment_date": {
      "format": "date",
      "title": "Payment Date",
      "type": "string"
     },
     "value_per_share": {
      "title": "Value Per Share",
      "type": "number"
     }
    },
    "required": [
     "payment_date",
     "value_per_share"
    ],
    "title": "FIIDividendResponse",
    "type": "object"
   },
   "FIIIndicatorsResponse": {
    "description": "The fund's own numbers, as of the report they came from.\n\nThe yields and the monthly return are ratios: 0.12381 is 12.381%.\n``price_to_nav`` is the published P/VP, a multiple around 1. None of them\nis scaled here; the client decides how each is written.",
    "properties": {
     "as_of_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of Date"
     },
     "dividend_yield_12m": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dividend Yield 12M"
     },
     "dividend_yield_1m": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dividend Yield 1M"
     },
     "equity": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Equity"
     },
     "monthly_return": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Monthly Return"
     },
     "nav_per_share": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nav Per Share"
     },
     "price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price"
     },
     "price_to_nav": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price To Nav"
     },
     "segment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Segment"
     },
     "segment_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Segment Type"
     },
     "shareholders": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shareholders"
     },
     "shares_outstanding": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Shares Outstanding"
     },
     "total_assets": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total Assets"
     }
    },
    "title": "FIIIndicatorsResponse",
    "type": "object"
   },
   "FIIMarketFundResponse": {
    "properties": {
     "administrator": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Administrator"
     },
     "asset_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Asset Id"
     },
     "cnpj": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cnpj"
     },
     "dividend_yield_12m": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dividend Yield 12M"
     },
     "investors": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Investors"
     },
     "management_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Management Type"
     },
     "mandate": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mandate"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "nav_per_share": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Nav Per Share"
     },
     "price": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price"
     },
     "price_to_nav": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price To Nav"
     },
     "segment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Segment"
     },
     "ticker": {
      "title": "Ticker",
      "type": "string"
     },
     "type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Type"
     }
    },
    "required": [
     "ticker",
     "name"
    ],
    "title": "FIIMarketFundResponse",
    "type": "object"
   },
   "FIIMarketResponse": {
    "properties": {
     "funds": {
      "items": {
       "$ref": "#/components/schemas/FIIMarketFundResponse"
      },
      "title": "Funds",
      "type": "array"
     },
     "source": {
      "title": "Source",
      "type": "string"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "funds",
     "total",
     "source"
    ],
    "title": "FIIMarketResponse",
    "type": "object"
   },
...
   "FIIProfileResponse": {
    "properties": {
     "dividends": {
      "default": [],
      "items": {
       "$ref": "#/components/schemas/FIIDividendResponse"
      },
      "title": "Dividends",
      "type": "array"
     },
     "indicators": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/FIIIndicatorsResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "ticker": {
      "title": "Ticker",
      "type": "string"
     }
    },
    "required": [
     "ticker"
    ],
    "title": "FIIProfileResponse",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "MarketDataSeriesHistoryPoint": {
    "properties": {
     "close": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Close"
     },
     "date": {
      "format": "date",
      "title": "Date",
      "type": "string"
     },
     "high": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "High"
     },
     "low": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Low"
     },
     "open": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Open"
     },
     "source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source"
     }
    },
    "required": [
     "date"
    ],
    "title": "MarketDataSeriesHistoryPoint",
    "type": "object"
   },
   "MarketDataSeriesOption": {
    "description": "Minimal identity used to pick a series in a selector.",
    "properties": {
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "short_name": {
      "title": "Short Name",
      "type": "string"
     },
     "symbol": {
      "title": "Symbol",
      "type": "string"
     }
    },
    "required": [
     "id",
     "short_name",
     "name",
     "symbol"
    ],
    "title": "MarketDataSeriesOption",
    "type": "object"
   },
   "MarketDataSeriesPoint": {
    "properties": {
     "date": {
      "format": "date",
      "title": "Date",
      "type": "string"
     },
     "value": {
      "title": "Value",
      "type": "number"
     }
    },
    "required": [
     "date",
     "value"
    ],
    "title": "MarketDataSeriesPoint",
    "type": "object"
   },
   "MarketDataSeriesResponse": {
    "description": "A registered series with everything that identifies and types it.",
    "properties": {
     "currency": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Currency"
       },
       {
        "type": "null"
       }
      ]
     },
     "currency_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Currency Id"
     },
     "frequency": {
      "title": "Frequency",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "series_type": {
      "title": "Series Type",
      "type": "string"
     },
     "short_name": {
      "title": "Short Name",
      "type": "string"
     },
     "symbol": {
      "title": "Symbol",
      "type": "string"
     },
     "value_type": {
      "title": "Value Type",
      "type": "string"
     }
    },
    "required": [
     "id",
     "short_name",
     "name",
     "symbol",
     "series_type",
     "value_type",
     "frequency"
    ],
    "title": "MarketDataSeriesResponse",
    "type": "object"
   },
   "MarketDataSeriesTimeSeries": {
    "additionalProperties": {
     "items": {
      "$ref": "#/components/schemas/MarketDataSeriesPoint"
     },
     "type": "array"
    },
    "description": "Each series by its short name, mapped to its observations.",
    "title": "MarketDataSeriesTimeSeries",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "auth/jwt/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "description": "API for portfolio consolidation",
  "title": "My Stonks API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/market_data/fii/market": {
   "get": {
    "description": "Complete FII catalogue with summarized BRAPI indicators.",
    "operationId": "get_fii_market_market_data_fii_market_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FIIMarketResponse"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Fii Market",
    "tags": [
     "FII"
    ]
   }
  },
  "/market_data/fii/{asset_id}/profile": {
   "get": {
    "description": "Indicators and distribution history published for one real-estate fund.",
    "operationId": "get_fii_profile_market_data_fii__asset_id__profile_get",
    "parameters": [
     {
      "in": "path",
      "name": "asset_id",
      "required": true,
      "schema": {
       "title": "Asset Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FIIProfileResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Fii Profile",
    "tags": [
     "FII"
    ]
   }
  },
...
  "/market_data/series": {
   "get": {
    "operationId": "list_market_data_series_market_data_series_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/MarketDataSeriesResponse"
         },
         "title": "Response List Market Data Series Market Data Series Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Market Data Series",
    "tags": [
     "Market Data Series"
    ]
   }
  },
  "/market_data/series/options": {
   "get": {
    "description": "Selectable series, for pickers that only need identity.",
    "operationId": "list_market_data_series_options_market_data_series_options_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/MarketDataSeriesOption"
         },
         "title": "Response List Market Data Series Options Market Data Series Options Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "List Market Data Series Options",
    "tags": [
     "Market Data Series"
    ]
   }
  },
  "/market_data/series/time_series": {
   "get": {
    "description": "Every series' history in one currency, keyed by short name, for charting.",
    "operationId": "get_series_time_series_market_data_series_time_series_get",
    "parameters": [
     {
      "in": "query",
      "name": "currency",
      "required": false,
      "schema": {
       "default": "BRL",
       "enum": [
        "BRL",
        "USD"
       ],
       "title": "Currency",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MarketDataSeriesTimeSeries"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Series Time Series",
    "tags": [
     "Market Data Series"
    ]
   }
  },
  "/market_data/series/{series_id}/history": {
   "get": {
    "operationId": "get_market_data_series_history_market_data_series__series_id__history_get",
    "parameters": [
     {
      "in": "path",
      "name": "series_id",
      "required": true,
      "schema": {
       "title": "Series Id",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "start_date",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start Date"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/MarketDataSeriesHistoryPoint"
         },
         "title": "Response Get Market Data Series History Market Data Series  Series Id  History Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Market Data Series History",
    "tags": [
     "Market Data Series"
    ]
   }
  },
...
 }
}